Add support for multiple data tables
Both the converter and the python script now have some sort of
support for handling all the tables in our AOD format.
For each ESD file we convert it into the Arrow tables,
attach some metadata to distinguish between them and finally
stream them through stdout. On the receiving side we
use the metadata to assign each table a mnemonic name.